separate static date utils from simulation
