+++ src/main/java/ldbc/snb/datagen/generator/generators/PersonGenerator.java
@@ -38,9 +38,9 @@ package ldbc.snb.datagen.generator.generators;
 import ldbc.snb.datagen.DatagenParams;
 import ldbc.snb.datagen.dictionary.Dictionaries;
 import ldbc.snb.datagen.entities.dynamic.person.Person;
+import ldbc.snb.datagen.util.DateUtils;
 import ldbc.snb.datagen.generator.distribution.DegreeDistribution;
 import ldbc.snb.datagen.generator.tools.PowerDistribution;
-import ldbc.snb.datagen.util.DateUtils;
 import ldbc.snb.datagen.util.LdbcConfiguration;
 import ldbc.snb.datagen.util.RandomGeneratorFarm;
