+++ src/main/java/ldbc/snb/datagen/serializer/InsertEventSerializer.java
@@ -46,9 +46,9 @@ import ldbc.snb.datagen.entities.dynamic.person.Person;
 import ldbc.snb.datagen.entities.dynamic.relations.ForumMembership;
 import ldbc.snb.datagen.entities.dynamic.relations.Knows;
 import ldbc.snb.datagen.entities.dynamic.relations.Like;
+import ldbc.snb.datagen.util.DateUtils;
 import ldbc.snb.datagen.hadoop.InsertEvent;
 import ldbc.snb.datagen.hadoop.key.updatekey.InsertEventKey;
-import ldbc.snb.datagen.util.DateUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.*;
 import org.apache.hadoop.io.SequenceFile;
